Tests/LibWeb/Text/expected/Editing/mouse-selection-release-outside-iframe.txt

selection after release: L
selection after unpressed move: L
